+++ src/core/CHANGELOG.md
@@ -3,6 +3,11 @@
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- Expand `class` in phoneme rules, both in input and output ([#6](https://gitlab.com/lexibook/lexibook/issues/6))
+- Update phonemes when applying rules
+
 ## [0.2.5] - 2020-01-08
 
 <!-- next-url -->